adding contracts for batch logic
